# backend/src/config/swagger.yaml
openapi: 3.0.1
info:
    title: Relational DB Designer
    description: Swagger API Docs
    license:
        name: MIT
        url: https://opensource.org/licenses/MIT
    version: 1.0.0
...
paths:
    /api/auth/signup:
        post:
            tags:
                - Authentication
            description: User sign up with email and password
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                password:
                                    type: string
            responses:
                400:
                    description: Email or password error
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: Email already used or password not strong enough
                200:
                    description: User created, return uid
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: user uid
    /api/auth/login:
        post:
            tags:
                - Authentication
            description: User login with email and password
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                password:
                                    type: string
            responses:
                400:
                    description: Email or password error
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: Email already used or password not strong enough
                200:
                    description: User logined, return uid
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: user uid
...
    /api/collab:
        get:
            tags:
                - Collaboration
            description: Gets the list of users that can access the ERD (ERid) or gets the list of ERD that a user (Uid) has access to. Only provide either ERid or Uid
            parameters:
                - in: query
                  name: ERid
                  schema:
                      type: string
                  description: ER Diagram ID
                - in: query
                  name: Uid
                  schema:
                      type: string
                  description: User ID
            responses:
                200:
                    description: Successfully gets the list of users that can access the ERD (ERid) or gets the list of ERD that a user (Uid) has access to
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: "[{'email': '123@gmail.com', 'permission': 'OWNER'}]"
                400:
                    description: Either Uid or ERid has to be defined as query
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: "Either Uid or ERid has to be defined as query"
                404:
                    description: ID does not exist
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: User does not exist
                501:
                    description: An error occured
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: An error occured.
        put:
            tags:
                - Collaboration
            description: Grants email specified permission for the ER diagram
            parameters:
                - in: query
                  name: owner
                  schema:
                      type: string
                  description: Owner of ER diagram
                - in: query
                  name: email
                  schema:
                      type: string
                  description: Email that owner wants to grant permission
                - in: query
                  name: ERid
                  schema:
                      type: string
                  description: ER Diagram ID
                - in: query
                  name: permission
                  schema:
                      type: string
                  description: Permission granted
            responses:
                400:
                    description: Permission should be READ or READ-WRITE
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: Permission should be READ or READ-WRITE
                403:
                    description: No permission
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: Only owner can change permissions
                404:
                    description: User with email does not exist
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: User does not exist
                501:
                    description: An error occured
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: An error occured.
                200:
                    description: Permission successfully granted
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: Permission successfully granted
    /api/collab/create-duplicate:
        post:
            tags:
                - Collaboration
            desciption: Create a duplicate of an ERD (ERid) for a user (Uid)
            parameters:
                - in: query
                  name: ERid
                  schema:
                      type: string
                  description: ER Diagram ID
                - in: query
                  name: Uid
                  schema:
                      type: string
                  description: User ID
            responses:
                400:
                    description: Uid and ERid have to be defined as queries.
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: Uid and ERid have to be defined as queries.
                403:
                    description: No permission
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: No permission to given ERid
                404:
                    description: ID does not exist
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: User does not exist
                501:
                    description: An error occured
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: An error occured.
                200:
                    description: Successfully duplicated ER diagram, returns new ERid
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: erid
...